adicionado crud para folha de pagamento

+++ Model/FolhaPagamneto.cs
@@ -8,10 +8,10 @@ namespace Model
         public int FuncionarioId { get; set; }
         public Funcionario funcionario { get; set; }
         public DateTime Data { get; set; }
-        public decimal Salario { get; set; }
-        public decimal Inss { get; set; }
-        public decimal Ir { get; set; }
-        public decimal SalarioLiquido { get; set; }
+        public double Salario { get; set; }
+        public double Inss { get; set; }
+        public double Ir { get; set; }
+        public double SalarioLiquido { get; set; }
